Synchronism the ch canbus code with master and v3.1_dev

+++ modules/canbus/vehicle/ch/protocol/turnsignal_command_113.cc
@@ -60,7 +60,7 @@ void Turnsignalcommand113::set_p_turn_signal_cmd(
   int x = turn_signal_cmd;
 
   Byte to_set(data + 0);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace ch
